Refactor to common TestModelStub
