+++ tests/Support/Models/Cart.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace Tipoff\Discounts\Tests\Support\Models;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Tipoff\Discounts\Contracts\DiscountableCart;
+use Tipoff\Support\Models\TestModelStub;
 
 class Cart extends Model implements DiscountableCart
 {
-    use HasFactory;
+    use TestModelStub;
 
     protected $guarded = [
         'id',
