+++ tests/Support/Models/Order.php
@@ -6,10 +6,11 @@ namespace Tipoff\Discounts\Tests\Support\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Tipoff\Support\Models\TestModelStub;
 
 class Order extends Model
 {
-    use HasFactory;
+    use TestModelStub;
 
     protected $guarded = [
         'id',
