+++ tests/Support/Models/User.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace Tipoff\Discounts\Tests\Support\Models;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
+use Tipoff\Support\Models\TestModelStub;
 
 class User extends Authenticatable
 {
-    use HasFactory;
+    use TestModelStub;
 
     protected $guarded = ['id'];
 }
